Drop TiledOpInterface implementations for tensor.extract_slice/insert_slice.
These tiled op interface was necessary for these ops when tile and
distribute were happening at the Flow level, so they arent necessary
anymore.
A consequence is that it is now possible to fold
`tensor.extract_slice/tensor.insert_slices` with
`flow.dispatch.tensor.load/flow.dispatch.tensor.store`
operations. Moving these patterns out of Codegen into Flow. These
folders arent made as canonicalizations for now since its unclear (to
me at least) if this involves implicit change to the representation of
parallel semantics. Such concerns have been raised previously. So as
an abundance of caution just using these patterns where they are
relevant, during dispatch region formation.
