+++ iree/compiler/Codegen/Common/BufferizeCopyOnlyDispatchesPass.cpp
@@ -29,97 +29,7 @@
 namespace mlir {
 namespace iree_compiler {
 
-/// Returns the `hal.interface.binding` a value comes from.
-static Optional<IREE::HAL::InterfaceBindingSubspanOp> getBindingSubspanOp(
-    Value v) {
-  Operation *definingOp = v.getDefiningOp();
-  if (!definingOp) return llvm::None;
-  if (auto interfaceOp =
-          dyn_cast<IREE::HAL::InterfaceBindingSubspanOp>(definingOp)) {
-    return interfaceOp;
-  }
-  if (auto loadOp = dyn_cast<IREE::Flow::DispatchTensorLoadOp>(definingOp)) {
-    return getBindingSubspanOp(loadOp.source());
-  }
-  return llvm::None;
-}
-
 namespace {
-
-/// Pattern to fold `flow.dispatch.tensor.load` -> `tensor.extract_slice`.
-// TODO(ravishankarm): Eventually this should go in as a canonicalization at the
-// Flow level.
-struct FoldTensorLoadWithExtractSlice
-    : OpRewritePattern<tensor::ExtractSliceOp> {
-  using OpRewritePattern<tensor::ExtractSliceOp>::OpRewritePattern;
-
-  LogicalResult matchAndRewrite(tensor::ExtractSliceOp extractSliceOp,
-                                PatternRewriter &rewriter) const override {
-    auto dispatchTensorLoadOp =
-        extractSliceOp.source()
-            .getDefiningOp<IREE::Flow::DispatchTensorLoadOp>();
-    if (!dispatchTensorLoadOp) return failure();
-
-    SmallVector<OpFoldResult> offsets, sizes, strides;
-    // `tensor.extract_slice` (i.e. the producer) folds **into**
-    // `flow.dispatch.tensor.load1 (i.e. the consumer).
-    if (failed(foldOffsetsSizesAndStrides(
-            rewriter, dispatchTensorLoadOp->getLoc(), dispatchTensorLoadOp,
-            extractSliceOp, dispatchTensorLoadOp.getDroppedDims(), offsets,
-            sizes, strides))) {
-      return failure();
-    }
-
-    rewriter.replaceOpWithNewOp<IREE::Flow::DispatchTensorLoadOp>(
-        extractSliceOp, extractSliceOp.getType(), dispatchTensorLoadOp.source(),
-        dispatchTensorLoadOp.source_dims(), offsets, sizes, strides);
-    return success();
-  }
-};
-
-/// Pattern to fold `tensor.insert_slice` with `flow.dispatch.tensor.store`
-/// oeprations.
-// TODO(ravishankarm): Eventually this should go in as a canonicalization at the
-// Flow level.
-struct FoldInsertSliceWithTensorStoreOp
-    : OpRewritePattern<IREE::Flow::DispatchTensorStoreOp> {
-  using OpRewritePattern<IREE::Flow::DispatchTensorStoreOp>::OpRewritePattern;
-
-  LogicalResult matchAndRewrite(
-      IREE::Flow::DispatchTensorStoreOp dispatchTensorStoreOp,
-      PatternRewriter &rewriter) const override {
-    auto insertSliceOp =
-        dispatchTensorStoreOp.value().getDefiningOp<tensor::InsertSliceOp>();
-    if (!insertSliceOp) return failure();
-
-    // Check that the `dest` of the `tensor.insert_slice` and target of the
-    // `flow.dispatch.tensor.store` are the same interface binding.
-    Optional<IREE::HAL::InterfaceBindingSubspanOp> destBinding =
-        getBindingSubspanOp(insertSliceOp.dest());
-    Optional<IREE::HAL::InterfaceBindingSubspanOp> targetBinding =
-        getBindingSubspanOp(dispatchTensorStoreOp.target());
-    if (!destBinding || !targetBinding ||
-        destBinding.getValue() != targetBinding.getValue()) {
-      return failure();
-    }
-
-    SmallVector<OpFoldResult> offsets, sizes, strides;
-    // `tensor.insert_slice` (i.e. the producer) folds **into**
-    // `flow.dispatch.tensor.store` (i.e. the consumer).
-    if (failed(foldOffsetsSizesAndStrides(
-            rewriter, dispatchTensorStoreOp->getLoc(), dispatchTensorStoreOp,
-            insertSliceOp, dispatchTensorStoreOp.getDroppedDims(), offsets,
-            sizes, strides))) {
-      return failure();
-    }
-
-    rewriter.replaceOpWithNewOp<IREE::Flow::DispatchTensorStoreOp>(
-        dispatchTensorStoreOp, insertSliceOp.source(),
-        dispatchTensorStoreOp.target(), dispatchTensorStoreOp.target_dims(),
-        offsets, sizes, strides);
-    return success();
-  }
-};
 
 /// Pass to bufferize early copy-only dispatches. This allows backends
 /// to use the `linalg.generic` operation generated for lowering the dispatch.
@@ -136,18 +46,7 @@ struct BufferizeCopyOnlyDispatchesPass
 }  // namespace
 
 void BufferizeCopyOnlyDispatchesPass::runOnOperation() {
-  MLIRContext *context = &getContext();
   ModuleOp module = getOperation();
-
-  /// First apply the `flow.dispatch.tensor.load` -> `tensor.extract_slice` and
-  /// `tensor.insert_slice` -> `flow.dispatch.tensor.store` patterns.
-  RewritePatternSet patterns(context);
-  patterns
-      .insert<FoldInsertSliceWithTensorStoreOp, FoldTensorLoadWithExtractSlice>(
-          context);
-  if (failed(applyPatternsAndFoldGreedily(module, std::move(patterns)))) {
-    return signalPassFailure();
-  }
 
   SmallVector<Operation *> copyOnlyFunctions;
   auto funcOps = module.getOps<func::FuncOp>();
